작업사항(22/1/4~5) - Gizmo(Transform, Rotation, Scale) 구현
Engine Client

1. Editor
GUI
GUI_Scene 클래스 수정
1) ImGuiGizmo를 사용하여 Gizmo기능 구현
2) 숫자 키패트 1: Transform, 2: Rotation, 3: Scale, 4: Universal(Transform + Rotation + Scale)
3) GameObject에 Rotation 조정이 있을 때 Transform Gizmo는 회전을 안 하는 오류가 있음 -> 추후에 개선

2. TODO
각 Component GUI 디테일 추가
Scene Save&Load 버그 수정
Editor 상태에 따라 Component의 값 수정여부 수정

Engine Library

1. Component
Transform 클래스 수정

2. TODO
Component 및 Resource Save&Load 완성하기
ResourceManager 리팩토링

+++ Project/DX11_2D_GameEngine_Lib/TileMap.cpp
@@ -78,7 +78,7 @@ void TileMap::Render()
 	auto transform = m_p_owner_game_object->GetComponent<Transform>();
 	if (m_tile_size != Vector2::Zero && m_tile_count != 0)
 	{
-		transform->SetMeshScale(m_tile_count_x * m_tile_size.x, m_tile_count_y * m_tile_size.y);
+		transform->SetMeshScale(static_cast<UINT>(m_tile_count_x * m_tile_size.x), static_cast<UINT>(m_tile_count_y * m_tile_size.y));
 
 		auto world_matrix = transform->GetWorldMatrix();
